refactor conversion code
